changing product selection page to use drop-down menu changing ComponentCalculation to use BigDecimal for calculations - will make this pluggable for different number-type implementations

+++ web-logic/src/main/java/org/lawrencebower/docgen/web_logic/business/component_calculation/Format.java
@@ -2,13 +2,14 @@ package org.lawrencebower.docgen.web_logic.business.component_calculation;
 
 import org.lawrencebower.docgen.core.exception.DocGenException;
 
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 
 public enum Format {
     CURRENCY,
     NONE;
 
-    public static String applyFormat(Format operator, Float operand) {
+    public static String applyFormat(Format operator, BigDecimal operand) {
 
         String returnVal = null;
